refactor: 스마트 메뉴 테스트, 타입  파일명 변경
