+++ client/src/components/smart-menu/SmallMenu.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import React, { FC } from 'react';
-import { IMenu } from 'types/types';
+import { IMenu } from 'types';
 
 interface ISmallMenuProps {
   menu: IMenu;
