+++ client/src/components/smart-menu/__test__/smart-menu.test.tsx
